derbskol_backend/requirements.txt
Django==5.2.5
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
Pillow==10.2.0
python-dotenv==1.0.1
django-cors-headers==4.3.1
